+++ hexital/core/candle_manager.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-from copy import deepcopy
 from datetime import datetime, timedelta
 from functools import cmp_to_key
 from typing import List, Optional, Set
@@ -45,7 +44,7 @@ class CandleManager:
         self.timeframe_fill = timeframe_fill
         self.candlestick = candlestick
 
-        self._tasks(True)
+        self._tasks()
 
     def __eq__(self, other) -> bool:
         if not isinstance(other, CandleManager):
@@ -70,10 +69,7 @@ class CandleManager:
     def name(self, name: str):
         self._name = name
 
-    def _tasks(self, to_sort: Optional[bool] = False):
-        if to_sort:
-            self.sort_candles()
-
+    def _tasks(self):
         self.collapse_candles()
         self.convert_candles()
         self.trim_candles()
@@ -91,9 +87,7 @@ class CandleManager:
             candles_.append(candles)
         elif isinstance(candles, dict):
             candles_.append(Candle.from_dict(candles))
-        elif isinstance(candles, list):
-            if not candles:
-                return
+        elif isinstance(candles, list) and candles:
             candle_ = candles[0]
             if isinstance(candle_, Candle):
                 candles_.extend(candles)
@@ -109,17 +103,18 @@ class CandleManager:
         self.sort_candles(candles_)
 
         to_sort = False
-        last_timestamp = self.candles[-1].timestamp if len(self.candles) > 0 else None
+        last_timestamp = self.candles[-1].timestamp if self.candles else None
 
         for candle in candles_:
             if last_timestamp and candle.timestamp < last_timestamp:
                 to_sort = True
-            if not candle.timeframe or not self.timeframe:
-                self.candles.append(deepcopy(candle))
-            elif candle.timeframe <= self.timeframe:
-                self.candles.append(deepcopy(candle))
+            if not candle.timeframe or not self.timeframe or candle.timeframe <= self.timeframe:
+                self.candles.append(candle.clean_copy())
 
-        self._tasks(to_sort)
+        if to_sort:
+            self.sort_candles()
+
+        self._tasks()
 
     def sort_candles(self, candles: Optional[List[Candle]] = None):
         """Sorts Candles in order of timestamp, accounts for collapsing"""
@@ -243,7 +238,9 @@ class CandleManager:
                     low=prev_candle.close,
                     volume=0,
                     timestamp=prev_candle.timestamp + timeframe,
+                    timeframe=prev_candle.timeframe,
                 )
+                fill_candle.aggregation_factor = 0
                 candles.insert(index, fill_candle)
 
             index += 1