+++ hexital/core/hexital.py
@@ -1,4 +1,4 @@
-from copy import copy, deepcopy
+from copy import copy
 from datetime import timedelta
 from importlib import import_module
 from typing import Dict, List, Optional, Set
@@ -28,6 +28,7 @@ class Hexital:
     _indicators: Dict[str, Indicator]
     description: Optional[str] = None
     _timeframe: Optional[timedelta] = None
+    _timeframe_name: str = DEFAULT_CANDLES
     timeframe_fill: bool = False
     candle_life: Optional[timedelta] = None
     candlestick: Optional[CandlestickType] = None
@@ -53,8 +54,11 @@ class Hexital:
         if candlestick:
             self.candlestick = validate_candlesticktype(candlestick)
 
+        if self._timeframe:
+            self._timeframe_name = timedelta_to_str(self._timeframe)
+
         self._candles = {
-            DEFAULT_CANDLES: CandleManager(
+            self._timeframe_name: CandleManager(
                 candles if isinstance(candles, list) else [],
                 candle_life=self.candle_life,
                 timeframe=self._timeframe,
@@ -62,13 +66,12 @@ class Hexital:
                 candlestick=self.candlestick,
             )
         }
-        self._candles[DEFAULT_CANDLES].name = DEFAULT_CANDLES
 
         self._indicators = self._validate_indicators(indicators) if indicators else {}
 
     @property
     def timeframe(self) -> str | None:
-        return timedelta_to_str(self._timeframe) if self._timeframe else None
+        return self._timeframe_name if self._timeframe else None
 
     @property
     def timeframes(self) -> Set[str]:
@@ -92,7 +95,7 @@ class Hexital:
 
     def candles(self, name: Optional[str | TimeFrame | timedelta | int] = None) -> List[Candle]:
         """Get a set of candles by using either a Timeframe or Indicator name"""
-        name_ = name if name else DEFAULT_CANDLES
+        name_ = name if name else self._timeframe_name
         timeframe_name = self._parse_timeframe(name)
 
         name_ = timeframe_name if timeframe_name else name_
@@ -118,7 +121,7 @@ class Hexital:
         """Attempts to retrieve a reading with a given Indicator name.
         `name` can use '.' to find nested reading, E.G `MACD_12_26_9.MACD`
         """
-        reading = reading_by_index(self._candles[DEFAULT_CANDLES].candles, name, index=index)
+        reading = reading_by_index(self._candles[self._timeframe_name].candles, name, index=index)
 
         if reading is not None:
             return reading
@@ -132,6 +135,11 @@ class Hexital:
 
     def prev_reading(self, name: str) -> float | dict | None:
         return self.reading(name, index=-2)
+
+    def readings(self) -> Dict[str, List[float | dict | None]]:
+        """Returns a Dictionary of all the Indicators and there results in a list format."""
+
+        return {name: indicator.as_list() for name, indicator in self._indicators.items()}
 
     def reading_as_list(self, name: str) -> List[float | dict | None]:
         """Find given indicator and returns the readings as a list
@@ -166,7 +174,7 @@ class Hexital:
         timeframe_name = self._parse_timeframe(timeframe)
 
         if timeframe_name and self._candles.get(timeframe_name):
-            self._candles[timeframe_name].append(deepcopy(candles))
+            self._candles[timeframe_name].append(candles)
         else:
             for candle_manager in self._candles.values():
                 candle_manager.append(candles)
@@ -204,8 +212,8 @@ class Hexital:
         if not timeframe:
             return None
 
-        if timeframe == DEFAULT_CANDLES:
-            return DEFAULT_CANDLES
+        if timeframe == self._timeframe_name:
+            return self._timeframe_name
 
         if not timeframe_validation(timeframe):
             return None
@@ -235,17 +243,18 @@ class Hexital:
 
         for indicator in valid_indicators.values():
             if not indicator.timeframe:
-                indicator.candle_manager = self._candles[DEFAULT_CANDLES]
+                indicator.candle_manager = self._candles[self._timeframe_name]
             elif indicator.timeframe and indicator.timeframe in self._candles:
                 indicator.candle_manager = self._candles[indicator.timeframe]
             else:
                 manager = CandleManager(
-                    deepcopy(self._candles[DEFAULT_CANDLES]).candles,
+                    [],
                     candle_life=self.candle_life,
                     timeframe=indicator._timeframe if indicator._timeframe else self._timeframe,
                     timeframe_fill=self.timeframe_fill,
                     candlestick=self.candlestick,
                 )
+                manager.append(self._candles[self._timeframe_name].candles)
                 self._candles[manager.name] = manager
                 indicator.candle_manager = self._candles[manager.name]
 
@@ -261,7 +270,9 @@ class Hexital:
             if indicator_class:
                 return indicator_class(**indicator)
             else:
-                raise InvalidIndicator(f"Indicator {indicator_name} does not exist. [{indicator}]")
+                raise InvalidIndicator(
+                    f"Indicator {indicator_name} does not exist. [{raw_indicator}]"
+                )
 
         elif indicator.get("analysis") and isinstance(indicator.get("analysis"), str):
             analysis_name = indicator.pop("analysis")
@@ -269,7 +280,7 @@ class Hexital:
 
             if not analysis_class:
                 raise InvalidAnalysis(
-                    f"analysis {analysis_name} does not exist in patterns or movements. [{indicator}]"
+                    f"analysis {analysis_name} does not exist in patterns or movements. [{raw_indicator}]"
                 )
 
             return Amorph(analysis=analysis_class, **indicator)