+++ hexital/indicators/vwap.py
@@ -8,6 +8,7 @@ from hexital.utils.timeframe import (
     TimeFrame,
     convert_timeframe_to_timedelta,
     round_down_timestamp,
+    timedelta_to_str,
     timeframe_validation,
 )
 
@@ -32,7 +33,7 @@ class VWAP(Indicator):
     anchor: Optional[str | TimeFrame | timedelta | int] = "D"
 
     def _generate_name(self) -> str:
-        return f"{self._name}_{self.anchor}"
+        return f"{self._name}_{timedelta_to_str(self.anchor)}"
 
     def _validate_fields(self):
         if not timeframe_validation(self.anchor):
@@ -41,7 +42,7 @@ class VWAP(Indicator):
         self.anchor = convert_timeframe_to_timedelta(self.anchor)
 
     def _initialise(self):
-        self.add_managed_indicator("VWAP_data", Managed(name=f"{self.name}_data"))
+        self.data = self.add_managed_indicator("VWAP_data", Managed(name=f"{self.name}_data"))
 
     def _calculate_reading(self, index: int) -> float | dict | None:
         candle = self.candles[index]
@@ -60,8 +61,6 @@ class VWAP(Indicator):
         pv = pv + (candle.volume * typical_price)
         vol = vol + candle.volume
 
-        self.managed_indicators["VWAP_data"].set_reading(
-            {"pv": pv, "vol": vol, "active_anchor": current_anchor}
-        )
+        self.data.set_reading({"pv": pv, "vol": vol, "active_anchor": current_anchor})
 
         return pv / vol