+++ tests/extra/profiling/common.py
@@ -3,6 +3,7 @@ import json
 import os
 import pstats
 from datetime import datetime
+from pathlib import Path
 
 PATH = "./tests/data/"
 PROFS = "./prof/"
@@ -17,10 +18,12 @@ def load_json_candles(name: str, path: str) -> list:
 
 
 class Profiler:
+    full_path: Path
+
     def __init__(self, name: str):
         self.name = name
-        if os.path.exists(PROFS) is False:
-            os.mkdir(PROFS)
+        self.full_path = Path(PROFS, os.environ.get("EXT_PATH", ""))
+        self.full_path.mkdir(parents=True, exist_ok=True)
 
     def __enter__(self):
         self.profiler = cProfile.Profile()
@@ -30,4 +33,4 @@ class Profiler:
     def __exit__(self, exception_type, exception_value, exception_traceback):
         self.profiler.disable()
         stats = pstats.Stats(self.profiler).sort_stats("cumtime")
-        stats.dump_stats(f"{PROFS}/{self.name}.prof")
+        stats.dump_stats(Path(self.full_path, f"{self.name}.prof"))